# docs/openapi/registration-login.yaml
openapi: 3.1.0
info:
  title: Monsters Registration and Login Contract
  version: 0.4.0
paths:
  /api/v1/auth/registration-policy:
    get:
      summary: Read the current required Terms and Privacy documents
      operationId: getRegistrationPolicy
      responses:
        "200":
          description: Current registration policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationPolicyEnvelope"
        "503":
          description: Required document configuration is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /api/v1/auth/register:
    post:
      summary: Accept an email registration without disclosing membership
      operationId: registerMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
      responses:
        "202":
          description: Registration request safely accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationAcceptedEnvelope"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "409":
          description: Accepted document versions are no longer current
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "429":
          description: Registration or resend rate limit reached
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitEnvelope"
        "503":
          description: Registration configuration is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /api/v1/auth/email-verification-requests:
    post:
      summary: Safely request another email verification message
      operationId: resendEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationResendRequest"
      responses:
        "202":
          description: Request accepted for known and unknown email alike
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailVerificationRequestAcceptedEnvelope"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "429":
          description: Cooldown or multidimensional rate limit reached
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitEnvelope"
        "503":
          description: Registration configuration is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /api/v1/auth/email-verifications:
    post:
      summary: Consume a single-use email verification token
      operationId: verifyMemberEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationRequest"
      responses:
        "200":
          description: Email verified and eligibility continuation issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailVerifiedEnvelope"
        "400":
          description: Token is invalid, used, revoked, or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /api/v1/auth/login:
    post:
      summary: Authenticate by verified Email or continue an incomplete member flow
      operationId: loginMemberByVerifiedEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifiedEmailLoginRequest"
      responses:
        "200":
          description: Authenticated session or purpose-limited continuation
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AuthenticatedLoginEnvelope"
                  - $ref: "#/components/schemas/ContinuationLoginEnvelope"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          description: Invalid credentials or a non-disclosable member state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /api/v1/auth/eligibility-policy:
    get:
      summary: Read the Taiwan eligibility and guardian document policy
      operationId: getEligibilityPolicy
      responses:
        "200":
          description: Current eligibility policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /api/v1/auth/eligibility-completions:
    post:
      summary: Complete eligibility with a purpose-limited continuation credential
      operationId: completeEligibility
      security:
        - continuationCredential: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EligibilityCompletionRequest"
      responses:
        "200":
          description: Eligibility classified or guardian consent requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Stable field validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          description: Continuation credential invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
  /api/v1/auth/guardian-consent-actions:
    post:
      summary: Inspect a short-lived single-use guardian action link
      operationId: inspectGuardianConsentAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianConsentTokenRequest"
      responses:
        "200": {description: Guardian action is valid}
        "400": {description: Link invalid, used, revoked, or expired}
  /api/v1/auth/guardian-consents:
    post:
      summary: Grant guardian consent using a 24-hour single-use link
      operationId: grantGuardianConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianConsentTokenRequest"
      responses:
        "200": {description: Guardian consent granted}
        "400": {description: Link invalid, used, revoked, or expired}
  /api/v1/auth/guardian-consent-withdrawal-requests:
    post:
      summary: Safely request a 15-minute guardian withdrawal link
      operationId: requestGuardianConsentWithdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianWithdrawalRequest"
      responses:
        "202": {description: Same response for matching and non-matching requests}
        "400": {description: Request shape invalid}
  /api/v1/auth/guardian-consent-withdrawals:
    post:
      summary: Withdraw guardian consent using a 15-minute single-use link
      operationId: withdrawGuardianConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianConsentTokenRequest"
      responses:
        "200": {description: Consent withdrawn and existing member access blocked}
        "400": {description: Link invalid, used, revoked, or expired}
  /api/auth/login:
    post:
      deprecated: true
      summary: Temporary migration endpoint accepting Email or legacy account
      operationId: loginLegacyMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyLoginRequest"
      responses:
        "200":
          description: Historical authenticated session or purpose-limited continuation
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/LegacyAuthenticatedLoginEnvelope"
                  - $ref: "#/components/schemas/ContinuationLoginEnvelope"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "401":
          description: Invalid credentials or a non-disclosable member state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
components:
  securitySchemes:
    continuationCredential:
      type: apiKey
      in: header
      name: Authorization
      description: "Authorization: Continuation <opaque credential>"
  schemas:
    ApiEnvelope:
      type: object
      required:
        - success
        - code
        - message
        - fieldErrors
        - requestId
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
        data: {}
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        requestId:
          type: string
          format: uuid
    RegistrationPolicyData:
      type: object
      additionalProperties: false
      required:
        - termsVersion
        - termsUrl
        - privacyVersion
        - privacyUrl
      properties:
        termsVersion:
          type: string
          minLength: 1
        termsUrl:
          type: string
          format: uri
          pattern: "^https://"
        privacyVersion:
          type: string
          minLength: 1
        privacyUrl:
          type: string
          format: uri
          pattern: "^https://"
    RegistrationPolicyEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: REGISTRATION_POLICY_AVAILABLE
            data:
              $ref: "#/components/schemas/RegistrationPolicyData"
    RegistrationRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - acceptedTermsVersion
        - acceptedPrivacyVersion
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 15
          maxLength: 128
          writeOnly: true
          description: >-
            Counted as Unicode code points after NFC normalization. The value is
            not trimmed and exact whole-password blocklist matches are rejected
            with VALIDATION_FAILED and a stable password field error key.
        acceptedTermsVersion:
          type: string
          minLength: 1
          maxLength: 80
        acceptedPrivacyVersion:
          type: string
          minLength: 1
          maxLength: 80
    RegistrationAcceptedEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: REGISTRATION_ACCEPTED
            data:
              type: "null"
    EmailVerificationResendRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 255
    EmailVerificationRequestAcceptedEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: EMAIL_VERIFICATION_REQUEST_ACCEPTED
            data:
              type: "null"
    EmailVerificationRequest:
      type: object
      additionalProperties: false
      required:
        - token
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 512
          writeOnly: true
          description: Raw 24-hour single-use token; the server stores only its SHA-256 hash.
    EmailVerifiedEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: EMAIL_VERIFIED
            data:
              $ref: "#/components/schemas/ContinuationLoginData"
    RateLimitData:
      type: object
      additionalProperties: false
      required:
        - retryAfter
      properties:
        retryAfter:
          type: integer
          format: int64
          minimum: 1
    RateLimitEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: RATE_LIMITED
            data:
              $ref: "#/components/schemas/RateLimitData"
    VerifiedEmailLoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          writeOnly: true
    LegacyLoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          maxLength: 255
          description: Temporary migration identifier accepting Email or legacy account.
        password:
          type: string
          format: password
          writeOnly: true
    AuthenticatedMember:
      type: object
      additionalProperties: false
      required:
        - publicId
        - email
        - userName
      properties:
        publicId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        userName:
          type:
            - string
            - "null"
    AuthenticatedLoginData:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          readOnly: true
        refreshToken:
          type: string
          readOnly: true
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        user:
          $ref: "#/components/schemas/AuthenticatedMember"
    LegacyAuthenticatedLoginData:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          readOnly: true
        refreshToken:
          type: string
          readOnly: true
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        user:
          type: object
          additionalProperties: true
    ContinuationNextAction:
      type: string
      enum:
        - VERIFY_EMAIL
        - COMPLETE_ELIGIBILITY
        - REACTIVATE_ACCOUNT
        - REVIEW_SUSPENSION
        - REVIEW_DELETION
    ContinuationLoginData:
      type: object
      additionalProperties: false
      required:
        - nextAction
        - continuationCredential
        - expiresIn
      properties:
        nextAction:
          $ref: "#/components/schemas/ContinuationNextAction"
        continuationCredential:
          type: string
          readOnly: true
          description: Ten-minute opaque purpose-limited credential returned once.
        expiresIn:
          type: integer
          format: int64
          const: 600
    AuthenticatedLoginEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: AUTHENTICATED
            data:
              $ref: "#/components/schemas/AuthenticatedLoginData"
    ContinuationLoginEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: AUTH_CONTINUATION_REQUIRED
            data:
              $ref: "#/components/schemas/ContinuationLoginData"
    LegacyAuthenticatedLoginEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              const: AUTHENTICATED
            data:
              $ref: "#/components/schemas/LegacyAuthenticatedLoginData"
    EligibilityCompletionRequest:
      type: object
      additionalProperties: false
      required: [serviceRegion, birthday, confirmPublicNicknameDisclosure]
      properties:
        serviceRegion: {type: string, pattern: "^[A-Za-z]{2}$"}
        birthday: {type: string, format: date}
        publicNickname:
          type: [string, "null"]
          maxLength: 120
          description: Normalized to NFC and validated as 2-30 Unicode code points by the server.
        guardianEmail: {type: [string, "null"], format: email, maxLength: 255}
        acceptedMinorNoticeVersion: {type: [string, "null"], maxLength: 80}
        guardianConsentVersion: {type: [string, "null"], maxLength: 80}
        confirmPublicNicknameDisclosure: {type: boolean}
        publicNicknameDisclosureVersion: {type: [string, "null"], maxLength: 80}
    GuardianConsentTokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: {type: string, minLength: 1, maxLength: 512, writeOnly: true}
    GuardianWithdrawalRequest:
      type: object
      additionalProperties: false
      required: [consentReference, guardianEmail]
      properties:
        consentReference: {type: string, format: uuid, maxLength: 36}
        guardianEmail: {type: string, format: email, maxLength: 255, writeOnly: true}